+++ types/node/ts3.1/tsconfig.json
@@ -9,7 +9,7 @@
         "module": "commonjs",
         "target": "es6",
         "lib": [
-            "es6",
+            "es5",
             "dom"
         ],
         "noImplicitAny": true,
