+++ types/node/tsconfig.json
@@ -2,7 +2,6 @@
     "files": [
         "index.d.ts",
         "common.d.ts",
-        "lib.d.ts",
         "inspector.d.ts",
         "node-tests.ts"
     ],
